% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-package.R
\docType{package}
\name{ftp-package}
\alias{ftp-package}
\alias{ftp}
\title{FTP client for R}
\description{
Provides methods for working with FTP servers. This package
doesn't act as a server, only a client.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
