% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-client.R
\name{ftp}
\alias{ftp}
\title{ftp client}
\arguments{
\item{path}{(character) a path to cache files in. required}

\item{type}{(character) type of cache. One of user_cache_dir (default),
user_log_dir, user_data_dir, user_config_dir, site_data_dir,
site_config_dir. required}
}
\description{
ftp client
}
\details{
\strong{Methods}
  \describe{
    \item{\code{pwd()}}{
      Print working directory
    }
    \item{\code{cd()}}{
      Change directory
    }
  }
}
\examples{
(x <- ftp("ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/"))
x$pwd()
x$list()

x$cd("daily")
x$pwd()
x$list()

# turn on active
x$set_pasv(FALSE)

# change port
# x$ftp_port(4567)
}
