% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-client.R
\name{ftp_client}
\alias{ftp_client}
\title{ftp client}
\arguments{
\item{path}{(character) a path to cache files in. required}

\item{type}{(character) type of cache. One of user_cache_dir (default),
user_log_dir, user_data_dir, user_config_dir, site_data_dir,
site_config_dir. required}
}
\description{
ftp client
}
\details{
\strong{Methods}
\describe{
\item{\code{pwd()}}{
Print working directory
}
\item{\code{cd()}}{
Change directory
}
\item{\code{ftp_port()}}{
Change ftp port
}
\item{\code{set_pasv()}}{
Change directory
}
\item{\code{get()}}{
Get a file from remote server
}
\item{\code{list()}}{
List files on remote server
}
}
}
\examples{
\dontrun{
(x <- ftp_client("ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/"))
x$pwd()
x$list()

x$cd("daily")
x$pwd()
x$list()

# turn on active
x$set_pasv(TRUE)
x

# change port
# x$ftp_port(4567)

# get a file
x$get('ghcnd-states.txt', disk = (f <- tempfile()))
readLines(f, n = 10)
close(file(f))

# put a file
x <- ftp_client("ftp://speedtest.tele2.net/")
x$cd('upload')
x$list()
f <- tempfile(fileext = ".txt")
writeLines("foo\nbar\n", con = f)
readLines(f)
res <- x$put(f, TRUE)
}
}
