% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_list.R
\name{ftp_list}
\alias{ftp_list}
\title{List files on an FTP server}
\usage{
ftp_list(url, just_list = FALSE, messages = FALSE, ...)
}
\arguments{
\item{url}{(character) URL for the FTP server}

\item{just_list}{(logical) list files only? Default: \code{FALSE}}

\item{messages}{(logical) verbose messages. Default: \code{FALSE}}

\item{...}{further args passed on to \pkg{curl}}
}
\value{
a tibble (i.e., a \code{data.frame})
}
\description{
List files on an FTP server
}
\examples{
\dontrun{
ftp_list("ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/")

# just list files
ftp_list("ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/", TRUE)

# more examples
ftp_list("ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/by_year/")
ftp_list("ftp://ftp.ncdc.noaa.gov/pub/data/noaa/2014/")
}
}
