% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_upload.R
\name{ftp_upload}
\alias{ftp_upload}
\title{Upload files to an FTP server}
\usage{
ftp_upload(url, path, messages = FALSE, ...)
}
\arguments{
\item{url}{(character) URL for the file on the FTP server}

\item{path}{(character) path}

\item{messages}{(logical) verbose messages. Default: \code{FALSE}}

\item{...}{further args passed on to \pkg{curl}}
}
\value{
string, path to the file on disk
}
\description{
Upload files to an FTP server
}
\examples{
\dontrun{
file <- tempfile(fileext = ".csv")
row.names(mtcars) <- NULL
write.csv(mtcars, file = file, row.names = FALSE)
read.csv(file)

(x <- ftp_upload(url = "demo.wftpserver.com", path = file))
cat(rawToChar(x$headers))
cat(rawToChar(x$content))


file <- tempfile(fileext = ".png")
row.names(mtcars) <- NULL
(x <- ftp_upload(url = "ftp://speedtest.tele2.net/upload", path = file))
cat(rawToChar(x$headers))
cat(rawToChar(x$content))
}
}
