% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files on an FTP server}
\usage{
list_files(url, ...)
}
\arguments{
\item{url}{URL for the FTP server}

\item{...}{further args passed on to \code{\link[RCurl]{getURL}}}
}
\description{
List files on an FTP server
}
\examples{
\dontrun{
url <- "ftp://localhost"
url <- "http://dd.meteo.gc.ca/"
url <- "ftp://ftp.ncdc.noaa.gov/pub/data/noaa/2014/"
list_files(url)
}
}

