% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_fetch.R
\name{ftp_fetch}
\alias{ftp_fetch}
\title{Fetch files from an FTP server}
\usage{
ftp_fetch(url, path, verbose = FALSE, ...)
}
\arguments{
\item{url}{(character) URL for the file on the FTP server}

\item{path}{(character) path}

\item{verbose}{(logical) verbose messages}

\item{...}{further args passed on to \pkg{curl}}
}
\value{
string, path to the file on disk
}
\description{
Fetch files from an FTP server
}
\examples{
\dontrun{
file <- tempfile(fileext = ".csv.gz")
ftp_fetch("ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/by_year/1882.csv.gz",
  path = file)
utils::read.delim(file, sep = ",", nrows = 10, header = FALSE)
}
}
