% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_upload.R
\name{ftp_upload}
\alias{ftp_upload}
\title{Upload files to an FTP server}
\usage{
ftp_upload(url, path, verbose = FALSE, ...)
}
\arguments{
\item{url}{(character) URL for the file on the FTP server}

\item{path}{(character) path}

\item{verbose}{(logical) verbose messages}

\item{...}{further args passed on to \pkg{curl}}
}
\value{
string, path to the file on disk
}
\description{
Upload files to an FTP server
}
\examples{
\dontrun{
file <- tempfile(fileext = ".csv.gz")
ftp_upload(url = "demo.wftpserver.com", path = file)
}
}
